fix: concurrency issue - WPB-21441
